Refactor services and news backoffice
