+++ src/pages/backoffice/news.css
@@ -29,7 +29,7 @@
 }
 
 .modalDelete{
-    position: absolute;
+    position: fixed;
     width: 100%;
     height: 100vh;
     top: 0;
