+++ src/services/categories.js
@@ -1,7 +1,17 @@
-import { useEffect, useState } from "react";
 import axios from "axios";
-const GetAllCategories = () => {
+import { useEffect, useState } from "react";
+
+export const deleteCategory = (id) => {
+    return axios.delete(`http://localhost:3000/categories/${id}`,{
+        headers: {
+            Authorization: `${localStorage.getItem("token")}`
+        }
+    });
+};
+
+export const GetAllCategories = () => {
     const token = localStorage.getItem("token");
+    const [refresh, setRefresh] = useState(false);
     const [categories, setCategories] = useState([]);
     const [loading, setLoading] = useState(false);
     const [error, setError] = useState(null);
@@ -20,12 +30,11 @@ const GetAllCategories = () => {
             } catch (error) {
                 setError(error);
             }
+            setRefresh(false);
             setLoading(false);
         };
         fetchData();
-    }, []);
+    }, [refresh]);
 
-    return { categories, loading, error};
+    return { categories, loading, error,setRefresh};
 }
-
-export default GetAllCategories;
