+++ src/services/news.js
@@ -1,6 +1,7 @@
 import { useEffect, useState } from "react";
 import axios from "axios";
-const GetAllNews = () => {
+
+export const GetAllNews = () => {
     const [news, setNews] = useState([]);
     const [loading, setLoading] = useState(false);
     const [error, setError] = useState(null);
@@ -24,4 +25,24 @@ const GetAllNews = () => {
     return { news, loading, error,setRefresh };
 }
 
-export default GetAllNews;
+export const GetNewsById = ({id}) => {
+    const [news, setNews] = useState([]);
+    const [loading, setLoading] = useState(false);
+    const [error, setError] = useState(null);
+
+    useEffect(() => {
+        const fetchData = async () => {
+            setLoading(true);
+            try {
+                const result = await axios.get(`http://localhost:3000/news/${id}`);
+                setNews(result.data.entry);
+            } catch (error) {
+                setError(error);
+            }
+            setLoading(false);
+        };
+        fetchData();
+    }, [id]);
+
+    return { news, loading, error };
+}
